requirements.txt
redis==3.2.1
Django==2.2.3
celery==4.3.0
PyYAML==5.1.1
psycopg2==2.7.4
